Introduce wallet sub-modules to regain control of codebase
